+++ src/main/java/edu/kis/vh/nursery/factory/DefaultRhymersFactory.java
@@ -1,29 +1,28 @@
 package edu.kis.vh.nursery.factory;
 
-import edu.kis.vh.nursery.defaultCountingOutRhymer;
+import edu.kis.vh.nursery.DefaultCountingOutRhymer;
 import edu.kis.vh.nursery.FIFORhymer;
 import edu.kis.vh.nursery.HanoiRhymer;
-import edu.kis.vh.nursery.factory.Rhymersfactory;
 
 public class DefaultRhymersFactory implements Rhymersfactory {
 
 	@Override
-	public defaultCountingOutRhymer GetStandardRhymer() {
-		return new defaultCountingOutRhymer();
+	public DefaultCountingOutRhymer GetStandardRhymer() {
+		return new DefaultCountingOutRhymer();
 	}
 
 	@Override
-	public defaultCountingOutRhymer GetFalseRhymer() {
-		return new defaultCountingOutRhymer();
+	public DefaultCountingOutRhymer GetFalseRhymer() {
+		return new DefaultCountingOutRhymer();
 	}
 
 	@Override
-	public defaultCountingOutRhymer GetFIFORhymer() {
+	public DefaultCountingOutRhymer GetFIFORhymer() {
 		return new FIFORhymer();
 	}
 
 	@Override
-	public defaultCountingOutRhymer GetHanoiRhymer() {
+	public DefaultCountingOutRhymer GetHanoiRhymer() {
 		return new HanoiRhymer();
 	}
